+++ src/Plugin.hpp
@@ -1,6 +1,8 @@
 namespace GOTHIC_NAMESPACE 
 {
-	bool EdgesFixAffectAllNpcs;
+	bool EdgesFixAffectPlayer;
+	bool EdgesFixAffectHumans;
+	bool EdgesFixAffectOtherMonsters;
 
 	int oCNpc::IsFlyingCreature_Union()
 	{
@@ -97,9 +99,31 @@ namespace GOTHIC_NAMESPACE
 		auto vob = static_cast<zCVob*>(This->m_pvClientData);
 		auto npc = zDYNAMIC_CAST<oCNpc>(vob);
 
-		if (!npc || npc == player || (!EdgesFixAffectAllNpcs && !npc->IsFlyingCreature_Union()))
+		if (!npc)
 		{
 			return;
+		}
+
+		if(npc == player)
+		{
+			if(!EdgesFixAffectPlayer)
+			{
+				return;
+			}
+		}
+		else if(!npc->IsFlyingCreature_Union() && npc->IsHuman())
+		{
+			if(!EdgesFixAffectHumans)
+			{
+				return;
+			}
+		}
+		else if(!npc->IsFlyingCreature_Union() && !npc->IsHuman())
+		{
+			if(!EdgesFixAffectOtherMonsters)
+			{
+				return;
+			}
 		}
 
 		int EndOffset = 0;
@@ -133,9 +157,31 @@ namespace GOTHIC_NAMESPACE
 		auto vob = static_cast<zCVob*>(This->m_pvClientData);
 		auto npc = zDYNAMIC_CAST<oCNpc>(vob);
 
-		if (!npc || npc == player || (!EdgesFixAffectAllNpcs && !npc->IsFlyingCreature_Union()))
+		if (!npc)
 		{
 			return;
+		}
+
+		if(npc == player)
+		{
+			if(!EdgesFixAffectPlayer)
+			{
+				return;
+			}
+		}
+		else if(!npc->IsFlyingCreature_Union() && npc->IsHuman())
+		{
+			if(!EdgesFixAffectHumans)
+			{
+				return;
+			}
+		}
+		else if(!npc->IsFlyingCreature_Union() && !npc->IsHuman())
+		{
+			if(!EdgesFixAffectOtherMonsters)
+			{
+				return;
+			}
 		}
 
 		int EndOffset = 0;
@@ -260,11 +306,13 @@ namespace GOTHIC_NAMESPACE
 
 	void Game_ApplySettings()
 	{
-		EdgesFixAffectAllNpcs = zoptions->ReadBool("FlyingCreaturesFix", "EdgesFixAffectAllNpcs", false);
+		EdgesFixAffectPlayer = zoptions->ReadBool("FlyingCreaturesFix", "EdgesFixAffectPlayer", false);
+		EdgesFixAffectHumans = zoptions->ReadBool("FlyingCreaturesFix", "EdgesFixAffectHumans", false);
+		EdgesFixAffectOtherMonsters = zoptions->ReadBool("FlyingCreaturesFix", "EdgesFixAffectOtherMonsters", false);
 	}
 
 	/*int __stdcall WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nShowCmd);
-	auto Hook_WinMain = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004F3E10, 0x00506810, 0x005000F0, 0x00502D70)), &WinMain, Union::HookType::Hook_Detours);
+	auto Hook_WinMain = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004F3E10, 0x00506810, 0x005000F0, 0x00502D70)), &WinMain, ::Union::HookType::Hook_Auto);
 	int __stdcall WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int nShowCmd)
 	{
 		Game_EntryPoint();
@@ -272,7 +320,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_Init(oCGame* self, void* vtable);
-	auto Hook_oCGame_Init = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x00636F50, 0x0065D480, 0x006646D0, 0x006C1060)), &oCGame_Init, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_Init = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x00636F50, 0x0065D480, 0x006646D0, 0x006C1060)), &oCGame_Init, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_Init(oCGame* self, void* vtable)
 	{
 		Hook_oCGame_Init(self, vtable);
@@ -280,7 +328,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall CGameManager_Done(CGameManager* self, void* vtable);
-	auto Hook_CGameManager_Done = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x00424850, 0x00427310, 0x004251A0, 0x004254E0)), &CGameManager_Done, Union::HookType::Hook_Detours);
+	auto Hook_CGameManager_Done = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x00424850, 0x00427310, 0x004251A0, 0x004254E0)), &CGameManager_Done, ::Union::HookType::Hook_Auto);
 	void __fastcall CGameManager_Done(CGameManager* self, void* vtable)
 	{
 		Game_Exit();
@@ -288,7 +336,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_Render(oCGame* self, void* vtable);
-	auto Hook_oCGame_Render = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063DBE0, 0x006648F0, 0x0066B930, 0x006C86A0)), &oCGame_Render, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_Render = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063DBE0, 0x006648F0, 0x0066B930, 0x006C86A0)), &oCGame_Render, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_Render(oCGame* self, void* vtable)
 	{
 		Game_PreLoop();
@@ -296,15 +344,15 @@ namespace GOTHIC_NAMESPACE
 		Game_PostLoop();
 	}*/
 
-	/*void __fastcall zCWorld_Render(Union::Registers& reg);
-	auto Partial_zCWorld_Render = Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x005F409C, 0x00614F4C, 0x0061A1F0, 0x00621940)), &zCWorld_Render);
-	void __fastcall zCWorld_Render(Union::Registers& reg)
+	/*void __fastcall zCWorld_Render(::Union::Registers& reg);
+	auto Partial_zCWorld_Render = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x005F409C, 0x00614F4C, 0x0061A1F0, 0x00621940)), &zCWorld_Render);
+	void __fastcall zCWorld_Render(::Union::Registers& reg)
 	{
 		Game_Loop();
 	}*/
 
 	/*void __fastcall zCMenu_Render(zCMenu* self, void* vtable);
-	auto Hook_zCMenu_Render = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004D0DA0, 0x004E14E0, 0x004DB270, 0x004DDC20)), &zCMenu_Render, Union::HookType::Hook_Detours);
+	auto Hook_zCMenu_Render = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004D0DA0, 0x004E14E0, 0x004DB270, 0x004DDC20)), &zCMenu_Render, ::Union::HookType::Hook_Auto);
 	void __fastcall zCMenu_Render(zCMenu* self, void* vtable)
 	{
 		Hook_zCMenu_Render(self, vtable);
@@ -312,7 +360,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_WriteSaveGame(oCGame* self, void* vtable, int slot, zBOOL saveGlobals);
-	auto Hook_oCGame_WriteSaveGame = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063AD80, 0x00661680, 0x006685D0, 0x006C5250)), &oCGame_WriteSaveGame, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_WriteSaveGame = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063AD80, 0x00661680, 0x006685D0, 0x006C5250)), &oCGame_WriteSaveGame, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_WriteSaveGame(oCGame* self, void* vtable, int slot, zBOOL saveGlobals)
 	{
 		Game_SaveBegin();
@@ -321,7 +369,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_LoadGame(oCGame* self, void* vtable, int slot, const zSTRING& levelPath);
-	auto Hook_oCGame_LoadGame = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063C070, 0x00662B20, 0x00669970, 0x006C65A0)), &oCGame_LoadGame, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_LoadGame = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063C070, 0x00662B20, 0x00669970, 0x006C65A0)), &oCGame_LoadGame, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_LoadGame(oCGame* self, void* vtable, int slot, const zSTRING& levelPath)
 	{
 		Game_LoadBegin_NewGame();
@@ -330,7 +378,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_LoadSaveGame(oCGame* self, void* vtable, int slot, zBOOL loadGlobals);
-	auto Hook_oCGame_LoadSaveGame = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063C2A0, 0x00662D60, 0x00669BA0, 0x006C67D0)), &oCGame_LoadSaveGame, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_LoadSaveGame = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063C2A0, 0x00662D60, 0x00669BA0, 0x006C67D0)), &oCGame_LoadSaveGame, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_LoadSaveGame(oCGame* self, void* vtable, int slot, zBOOL loadGlobals)
 	{
 		Game_LoadBegin_SaveGame();
@@ -339,7 +387,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_ChangeLevel(oCGame* self, void* vtable, const zSTRING& levelpath, const zSTRING& startpoint);
-	auto Hook_Game_Load_ChangeLevel = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063CD60, 0x00663950, 0x0066A660, 0x006C7290)), &oCGame_ChangeLevel, Union::HookType::Hook_Detours);
+	auto Hook_Game_Load_ChangeLevel = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063CD60, 0x00663950, 0x0066A660, 0x006C7290)), &oCGame_ChangeLevel, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_ChangeLevel(oCGame* self, void* vtable, const zSTRING& levelpath, const zSTRING& startpoint)
 	{
 		Game_LoadBegin_ChangeLevel();
@@ -348,7 +396,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_TriggerChangeLevel(oCGame* self, void* vtable, const zSTRING& levelpath, const zSTRING& startpoint);
-	auto Hook_oCGame_TriggerChangeLevel = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063D480, 0x00664100, 0x0066AD80, 0x006C7AF0)), &oCGame_TriggerChangeLevel, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_TriggerChangeLevel = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063D480, 0x00664100, 0x0066AD80, 0x006C7AF0)), &oCGame_TriggerChangeLevel, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_TriggerChangeLevel(oCGame* self, void* vtable, const zSTRING& levelpath, const zSTRING& startpoint)
 	{
 		Game_LoadBegin_TriggerChangeLevel();
@@ -358,7 +406,7 @@ namespace GOTHIC_NAMESPACE
 
 /*#if ENGINE <= Engine_G1A
 	void __fastcall oCGame_Pause_G1(oCGame* self, void* vtable);
-	auto Hook_oCGame_Pause = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063DF50, 0x00664CD0, 0, 0)), &oCGame_Pause_G1, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_Pause = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063DF50, 0x00664CD0, 0, 0)), &oCGame_Pause_G1, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_Pause_G1(oCGame* self, void* vtable)
 	{
 		Hook_oCGame_Pause(self, vtable);
@@ -366,7 +414,7 @@ namespace GOTHIC_NAMESPACE
 	}
 #else
 	void __fastcall oCGame_Pause_G2(oCGame* self, void* vtable, zBOOL sessionPaused);
-	auto Hook_oCGame_Pause = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0, 0, 0x0066BD50, 0x006C8AD0)), &oCGame_Pause_G2, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_Pause = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0, 0, 0x0066BD50, 0x006C8AD0)), &oCGame_Pause_G2, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_Pause_G2(oCGame* self, void* vtable, zBOOL sessionPaused)
 	{
 		Hook_oCGame_Pause(self, vtable, sessionPaused);
@@ -375,7 +423,7 @@ namespace GOTHIC_NAMESPACE
 #endif*/
 
 	/*void __fastcall oCGame_Unpause(oCGame* self, void* vtable);
-	auto Hook_oCGame_Unpause = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063E1D0, 0x00664F80, 0x0066BFD0, 0x006C8D50)), &oCGame_Unpause, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_Unpause = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x0063E1D0, 0x00664F80, 0x0066BFD0, 0x006C8D50)), &oCGame_Unpause, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_Unpause(oCGame* self, void* vtable)
 	{
 		Hook_oCGame_Unpause(self, vtable);
@@ -383,7 +431,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	/*void __fastcall oCGame_DefineExternals_Ulfi(oCGame* self, void* vtable, zCParser* parser);
-	auto Hook_oCGame_DefineExternals_Ulfi = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x006495B0, 0x006715F0, 0x00677A00, 0x006D4780)), &oCGame_DefineExternals_Ulfi, Union::HookType::Hook_Detours);
+	auto Hook_oCGame_DefineExternals_Ulfi = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x006495B0, 0x006715F0, 0x00677A00, 0x006D4780)), &oCGame_DefineExternals_Ulfi, ::Union::HookType::Hook_Auto);
 	void __fastcall oCGame_DefineExternals_Ulfi(oCGame* self, void* vtable, zCParser* parser)
 	{
 		Hook_oCGame_DefineExternals_Ulfi(self, vtable, parser);
@@ -391,7 +439,7 @@ namespace GOTHIC_NAMESPACE
 	}*/
 
 	void __fastcall CGameManager_ApplySomeSettings(CGameManager* self, void* vtable);
-	auto Hook_CGameManager_ApplySomeSettings = Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004267C0, 0x004291E0, 0x00427370, 0x004276B0)), &CGameManager_ApplySomeSettings, Union::HookType::Hook_Detours);
+	auto Hook_CGameManager_ApplySomeSettings = ::Union::CreateHook(reinterpret_cast<void*>(zSwitch(0x004267C0, 0x004291E0, 0x00427370, 0x004276B0)), &CGameManager_ApplySomeSettings, ::Union::HookType::Hook_Auto);
 	void __fastcall CGameManager_ApplySomeSettings(CGameManager* self, void* vtable)
 	{
 		Hook_CGameManager_ApplySomeSettings(self, vtable);